+++ store/app/src/main/java/br/com/stone/vianna/starstore/di/Modules.kt
@@ -13,6 +13,7 @@ import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionContr
 import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionPresenter
 import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionRepository
 import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionRepositoryImpl
+import br.com.stone.vianna.starstore.helper.StringResolver
 import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
@@ -28,6 +29,7 @@ val applicationModule = module(override = true) {
 
     single { createWebService<ItemListDataSource>(get(), BuildConfig.ITEMS_API) }
     single { createWebService<PaymentDataSource>(get(), BuildConfig.PAYMENT_API) }
+
 }
 
 val repositoryModule = module {
