+++ store/app/src/main/java/br/com/stone/vianna/starstore/entity/ItemDao.kt
@@ -1,6 +1,7 @@
 package br.com.stone.vianna.starstore.entity
 
 import android.arch.persistence.room.*
+import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
 
@@ -8,13 +9,13 @@ import io.reactivex.Single
 interface ItemDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insertItem(item: Item)
+    fun insertItem(item: Item): Completable
 
     @Update
-    fun updateItem(item: Item)
+    fun updateItem(item: Item): Completable
 
     @Delete
-    fun deleteItem(item: Item)
+    fun deleteItem(item: Item): Completable
 
     @Query("SELECT * FROM Item WHERE title == :title")
     fun getItemByName(title: String): Single<List<Item>>
@@ -26,5 +27,5 @@ interface ItemDao {
     fun getItemsCount(): Maybe<Int>
 
     @Query("DELETE FROM Item")
-    fun removeItems()
+    fun removeItems(): Completable
 }
