+++ store/app/src/main/java/br/com/stone/vianna/starstore/entity/PaymentTransaction.kt
@@ -1,6 +1,5 @@
 package br.com.stone.vianna.starstore.entity
 
-import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
 
@@ -9,12 +8,8 @@ class PaymentTransaction {
 
     @PrimaryKey(autoGenerate = true)
     var id: Int? = null
-    @ColumnInfo(name = "cardHolderName")
     var cardHolderName = ""
-    @ColumnInfo(name = "cardLastDigits")
     var cardLastDigits = ""
-    @ColumnInfo(name = "value")
     var value = 0
-    @ColumnInfo(name = "transactionDateTime")
     var transactionDateTime = ""
 }
