+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/card/CreditCardActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import br.com.stone.vianna.starstore.R
 import br.com.stone.vianna.starstore.baseClasses.BaseActivity
 import br.com.stone.vianna.starstore.entity.PaymentRequest
-import br.com.stone.vianna.starstore.extensions.hide
-import br.com.stone.vianna.starstore.extensions.show
+import br.com.stone.vianna.starstore.helper.hide
+import br.com.stone.vianna.starstore.helper.show
 import kotlinx.android.synthetic.main.activity_credit_card.*
 import org.koin.android.ext.android.inject
 import org.koin.core.parameter.parametersOf
