+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/transactionHistory/TransactionAdapter.kt
@@ -4,8 +4,8 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import  br.com.stone.vianna.starstore.extensions.convertDateToFormat
-import  br.com.stone.vianna.starstore.extensions.toMoneyFormat
+import  br.com.stone.vianna.starstore.helper.convertDateToFormat
+import  br.com.stone.vianna.starstore.helper.toMoneyFormat
 import  br.com.stone.vianna.starstore.R
 import  br.com.stone.vianna.starstore.entity.PaymentTransaction
 import kotlinx.android.synthetic.main.view_item_transaction.view.*
