+++ store/app/src/main/java/br/com/stone/vianna/starstore/helper/Extensions.kt
@@ -1,14 +1,9 @@
-package br.com.stone.vianna.starstore.extensions
+package br.com.stone.vianna.starstore.helper
 
 import android.view.View
-import br.com.stone.vianna.starstore.R
-import br.com.stone.vianna.starstore.StoreApplication
-import com.google.gson.Gson
-import com.jakewharton.retrofit2.adapter.rxjava2.HttpException
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import java.io.IOException
 import java.text.NumberFormat
 import java.text.SimpleDateFormat
 import java.util.*
@@ -34,39 +29,6 @@ fun <T> Observable<T>.addThreads(): Observable<T> {
 }
 
 /**
- * Handle Error
- */
-
-data class ResponseParser(val success: Boolean, val error: String, val code: Int = 404)
-
-val Throwable.parser: ResponseParser
-    get() {
-        if (this !is HttpException) {
-            return ResponseParser(false, R.string.error_title.translate)
-        }
-
-        val strError = this.response().errorBody()?.string()
-                ?: return ResponseParser(false, R.string.error_title.translate, this.response().code())
-
-        val gson = Gson()
-        val typeAdapter = gson.getAdapter(ResponseParser::class.java)
-        return try {
-            typeAdapter.fromJson(strError)
-        } catch (e: IOException) {
-            ResponseParser(false, R.string.error_title.translate, this.response().code())
-        }
-    }
-
-/**
- * Translate
- */
-
-val Int.translate: String
-    get() {
-        return StoreApplication.context.getString(this) ?: ""
-    }
-
-/**
  * Money
  */
 fun Int.toMoneyFormat(): String {
@@ -75,7 +37,6 @@ fun Int.toMoneyFormat(): String {
     val value = currency.format(this / 100.0)
     return value
 }
-
 
 /**
  * Form Helper
@@ -107,8 +68,8 @@ fun validateCardExpiryDate(expiryDate: String): Boolean {
 }
 
 /**
-* Date
-*/
+ * Date
+ */
 
 fun convertDateToFormat(date: String, oldFormat: String, newFormat: String): String {
     val initialFormat = SimpleDateFormat(oldFormat)
